pyproject.toml
[project]
name = "discordbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "google>=3.0.0",
    "google-genai>=1.68.0",
    "google-generativeai>=0.8.6",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]
